asterisk/wazo-res-stasis-amqp/integration_tests/test-requirements.txt
git+https://github.com/rsichny/swagger-py.git  # the digium release does not support python3
https://github.com/wazo-platform/ari-py/archive/0.4.0.zip
https://github.com/wazo-platform/xivo-test-helpers/archive/master.zip
docker-compose
kombu
pyhamcrest
pytest
